- Add the main page of the site.
This is just a template for the actuall index.html that is generated.
Currently only the generate_post_list_for_index.py is being used to
create a real index.html.

In the future if is needed we need to change the output of the script
to actually output just plain html that will be merged by a possible
future script ;D